finish permissions section of course

+++ sick-fits/backend/src/resolvers/Mutation.js
@@ -3,6 +3,7 @@ const jwt = require('jsonwebtoken');
 const { randomBytes } = require('crypto');
 const { promisify } = require('util');
 const { transport, makeANiceEmail } = require('../mail');
+const { hasPermission } = require('../utils');
 
 const Mutations = {
     async createItem(parent, args, ctx, info) {
@@ -42,9 +43,13 @@ const Mutations = {
     async deleteItem(parent, args, ctx, info) {
         const where = { id: args.id };
         //find the item
-        const item = await ctx.db.query.item({ where }, `{ id title }`);
+        const item = await ctx.db.query.item({ where }, `{ id title user { id } }`);
         //check if they own that item
-        //To do
+        const ownsItem = item.user.id === ctx.request.userId;
+        const hasPermissions = ctx.request.user.permissions.some(permission => ['ADMIN', 'ITEMDELETE'].includes(permission))
+        if(!ownsItem && !hasPermissions) {
+            throw new Error("You don't have permission to do that!");
+        }
         //delete it
         return ctx.db.mutation.deleteItem({ where }, info);
     },
@@ -160,7 +165,36 @@ const Mutations = {
         });
         //return new user
         return updatedUser;
-    }
+    },
+    async updatePermissions(parent, args, ctx, info) {
+        //check if logged in
+        if(!ctx.request.userId) {
+            throw new Error('You must be logged in!');
+        }
+        //query the current user
+        const currentUser = await ctx.db.query.user(
+            {
+                where: {
+                    id: ctx.request.userId,
+                },
+            },
+            info
+        );
+        //check if they have permissions to do this
+        hasPermission(currentUser, ['ADMIN', 'PERMISSIONUPDATE']);
+        //update the permissions
+        return ctx.db.mutation.updateUser({
+            data: {
+                permissions: {
+                    set: args.permissions,
+                },
+            },
+            where: {
+                id: args.userId
+            },
+        },
+        info
+    )}
 };
 
 module.exports = Mutations;
